Warnings
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@8278 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/int/distinct/dom.hpp
@@ -321,7 +321,7 @@ namespace Gecode { namespace Int { namespace Distinct {
       max = std::max(max,x[i].max());
     }
 
-    unsigned int width = max-min+1;
+    unsigned int width = static_cast<unsigned int>(max-min+1);
 
     // Definitly not enough values
     if (width < static_cast<unsigned int>(n_view))
